Added two combateffects: Shoot and Heal

+++ entities/entity.h
@@ -33,7 +33,7 @@ public:
 
 	EntityStats& stats;
 	Vector2d speed;
-	double HP;
+	int HP;
 
 	Entity(World* world, Vector2d newPos, EntityStats& stats);
 	bool isInArea(Vector2d px1, Vector2d px2);
